requirements.txt
python-dotenv
google-generativeai
langchain_google_genai
langchain
PyPDF2
faiss-cpu
streamlit
-e .